add association and validation to the car model

+++ db/migrate/20230104190213_create_cars.rb
@@ -4,7 +4,7 @@ class CreateCars < ActiveRecord::Migration[7.0]
       t.string :name
       t.string :image
       t.string :model
-      t.integer :daily_price
+      t.float :daily_price
       t.text :description 
       t.timestamps
     end
